+++ src/helpers/file.helper.ts
@@ -2,8 +2,8 @@ import * as vscode from "vscode";
 import * as fs from "fs";
 import * as path from "path";
 import { MessageStatus, Organization } from "../constants/index";
-import { Logger } from "./loggerHelper";
-import { DebrickedDataHelper } from "./debrickedDataHelper";
+import { Logger } from "./logger.helper";
+import { DebrickedDataHelper } from "./debrickedData.helper";
 
 export class FileHelper {
     constructor(
