+++ src/helpers/statusBarMessage.helper.ts
